Extract projection terrain height sampler
